service/requirements.txt
dataclasses
tornado
numpy
pandas
sklearn
catboost
torch
torchvision
tqdm
pyocr
opencv-python
Pillow
pdf2image
imutils
ezdxf
shapely
cairosvg
jsonschema
typeguard
webargs
yamlparams